# backend2/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BBWS Sumatera VIII API Reference",
    "version": "1.0.0",
    "description": "Dokumentasi REST API BBWS Sumatera VIII Hono Engine (Auto-generated)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server Lokal"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/balai/bbwssumatera8/api/berita": {
      "get": {
        "tags": [
          "Berita"
        ],
        "summary": "GET /balai/bbwssumatera8/api/berita",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      },
      "post": {
        "tags": [
          "Berita"
        ],
        "summary": "POST /balai/bbwssumatera8/api/berita",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/berita/highlighted": {
      "get": {
        "tags": [
          "Berita"
        ],
        "summary": "GET /balai/bbwssumatera8/api/berita/highlighted",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      }
    },
    "/balai/bbwssumatera8/api/berita/{id}": {
      "get": {
        "tags": [
          "Berita"
        ],
        "summary": "GET /balai/bbwssumatera8/api/berita/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      },
      "put": {
        "tags": [
          "Berita"
        ],
        "summary": "PUT /balai/bbwssumatera8/api/berita/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Berita"
        ],
        "summary": "DELETE /balai/bbwssumatera8/api/berita/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/banner": {
      "get": {
        "tags": [
          "Banner"
        ],
        "summary": "GET /balai/bbwssumatera8/api/banner",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      }
    },
    "/balai/bbwssumatera8/api/banner/upload": {
      "post": {
        "tags": [
          "Banner"
        ],
        "summary": "POST /balai/bbwssumatera8/api/banner/upload",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/banner/{id}": {
      "delete": {
        "tags": [
          "Banner"
        ],
        "summary": "DELETE /balai/bbwssumatera8/api/banner/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}